Converter: VB6 data types and runtime aligned with VB Migration Partner
- Widths as VBMP: Integer -> short, Long -> int, Single -> float, Double -> double,
  Currency -> decimal; Double literals; Null -> DBNull.Value
- Implicit conversions VB6 applies on assignment and argument passing via
  Microsoft.VisualBasic Conversions (rounding, parsing), typed from the expression;
  "/" is a floating division; Currency mixed with Single/Double is promoted;
  CInt/CLng/CStr/... map to Conversions (CInt is 16-bit)
- Numeric / Variant conditions tested against zero; Not on numbers stays bitwise
- Arrays are T[] (Split-compatible, UBound/LBound work); non-zero lower bounds and
  Option Base 1 use VB6Array<T> (ref indexer); ReDim/Erase for both; module-level
  variables known inside procedures (array indexing, types)
- Type -> struct : IVbStruct with Initialize (strings, fixed strings, fixed arrays,
  nested UDTs) and MarshalAs for API calls; UDT variables and arrays initialized
- String * n: locals padded on assignment, fields through a FixedLen property
- Module-level As New: auto-instancing property; DefType applied to declarations
- UpgradeHelpers: VbRuntime (NewArray, ReDim, NewStruct, FixedLen, MidStmt,
  TextCompare) and VB6Array<T>; generated project references the package (net48)

+++ Vb6ToCSharp.Tests/ConfigTests.cs
@@ -29,10 +29,10 @@ public class ConfigTests : IClassFixture<ConverterFixture>
         WithSection(ModConfig.iniSectionDataTypes, () =>
         {
             Assert.Equal("MyCs", ModVb6ToCs.ConvertDataType("MyType"));
-            Assert.Equal("double", ModVb6ToCs.ConvertDataType("Double"));
+            Assert.Equal("decimal", ModVb6ToCs.ConvertDataType("Double"));
             Assert.Equal("int", ModVb6ToCs.ConvertDataType("Long"));
-        }, ("MyType", "MyCs"), ("Double", "double"));
-        Assert.Equal("decimal", ModVb6ToCs.ConvertDataType("Double"));
+        }, ("MyType", "MyCs"), ("Double", "decimal"));
+        Assert.Equal("double", ModVb6ToCs.ConvertDataType("Double")); // VB6 Double is a binary double
     }
 
     [Fact]
